Rename ascwds worker cleaned column names file
