+++ utils/validation/validation_rules/ascwds_worker_cleaned_validation_rules.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from utils.column_names.cleaned_data_files.ascwds_worker_cleaned_values import (
+from utils.column_names.cleaned_data_files.ascwds_worker_cleaned import (
     AscwdsWorkerCleanedColumns as AWKClean,
 )
 from utils.column_values.categorical_columns_by_dataset import (
